+++ src/main/java/inf/unideb/caloriecounterbackend/service/ProductService.java
@@ -4,6 +4,7 @@ import inf.unideb.caloriecounterbackend.dto.ProductDTO;
 import inf.unideb.caloriecounterbackend.dto.Result;
 import inf.unideb.caloriecounterbackend.entity.Product;
 import inf.unideb.caloriecounterbackend.exception.ApplicationError;
+import inf.unideb.caloriecounterbackend.exception.ApplicationException;
 import inf.unideb.caloriecounterbackend.repository.ProductRepository;
 
 import java.util.List;
@@ -28,6 +29,7 @@ public class ProductService extends BaseService<ProductDTO, Product> {
     public Result<ProductDTO> createProduct(final ProductDTO productDTO) {
         final Product product = super.mapFromDTO(productDTO);
         product.setId(null);
+        product.setUserId(super.getUserUuid());
         return new Result<>(super.mapToDTO(this.productRepository.save(product)));
     }
 
@@ -49,6 +51,10 @@ public class ProductService extends BaseService<ProductDTO, Product> {
     public Result<ProductDTO> updateProduct(final ProductDTO productDTO, final String productId) {
         return this.productRepository.findById(productId)
                 .map(updatedProduct -> {
+                    if (!updatedProduct.getUserId().equals(super.getUserUuid())) {
+                        throw new ApplicationException(ApplicationError.notMatchingUserId());
+                    }
+                    productDTO.setUserId(updatedProduct.getUserId());
                     updatedProduct = super.mapFromDTO(productDTO);
                     updatedProduct.setId(productId);
                     return super.mapToDTO(this.productRepository.save(updatedProduct));
@@ -58,8 +64,12 @@ public class ProductService extends BaseService<ProductDTO, Product> {
     }
 
     public Result<Void> deleteProduct(final String productId) {
-        this.productRepository.deleteById(productId);
-        return Result.ok();
+        if (this.productRepository.findById(productId).orElseThrow().getUserId().equals(super.getUserUuid())) {
+            this.productRepository.deleteById(productId);
+            return Result.ok();
+        } else {
+            return Result.error(ApplicationError.notMatchingUserId());
+        }
     }
 
 }
